requirements.txt
appnope==0.1.0
astroid==1.5.3
autopep8==1.3.3
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
beautifulsoup4==4.6.3
biopython==1.73
bleach==1.5.0
boost==0.1
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
chromedriver==2.24.1
colorama==0.3.9
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.1
isort==4.2.15
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Keras==2.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.5.0
Markdown==2.6.9
MarkupSafe==1.0
Mastodon.py==1.1.1
matplotlib==2.0.2
mccabe==0.6.1
mercurial==4.5.3
mistune==0.7.4
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.1.0
numpy==1.13.3
pandas==0.21.0
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==5.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.2
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pycodestyle==2.3.1
Pygments==2.2.0
pylint==1.7.4
pyparsing==2.2.0
python-config==0.1.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
scandir==1.5
scikit-learn==0.19.1
scipy==1.0.0
selenium==3.14.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
SQLAlchemy==1.1.15
subprocess32==3.2.7
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.3
wrapt==1.10.11
